pyproject.toml
[tool.poetry]
name = "prynterface"
version = "0.0.1"
description = "Application for interfacing with 3D printers over serial connection."
authors = ["Dominik Ullrich <dominik119017@gmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.7.0"
pyserial = "^3.5"
pyserial-asyncio = "^0.6"
customtkinter = "^5.1.2"
numpy = "^1.24.2"
matplotlib = "^3.7.1"
aiostream = "^0.4.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
coverage = "^7.2.3"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.21.0"
pylint = "^2.17.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
